requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
python-dotenv==1.0.1
PyMySQL==1.1.1
SQLAlchemy==2.0.32
reportlab==4.2.2
Werkzeug==3.0.3
requests==2.32.3
